fix : change testcode dir and add all failed case
